requirements.txt
requests
openai
pandas
tabulate
matplotlib
opencv-python
discord.py[voice]>=2.3.0
PyNaCl>=1.5.0
google-cloud-texttospeech
python-dotenv>=1.0.0